update detail screen

+++ app/src/main/java/com/tdpc/toko99/core/data/remote/response/SatuanBarang.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class SatuanBarang(
 
     @field:SerializedName("harga")
-    val harga: Int? = null,
+    val harga: Double? = null,
 
     @field:SerializedName("updated_at")
     val updatedAt: String? = null,
